Connect Timeline to contentful
