+++ components/Timeline.jsx
@@ -1,13 +1,12 @@
 /* eslint-disable @next/next/no-img-element */
 import Image from 'next/image';
-import { TIMELINE_DATA } from '../data/timeline';
 import { UnderlineTypes } from '../utils/underlineType';
 import Heading from './Heading';
-
+import { rgbDataURL } from '../utils/blurImage';
 const TIMELINE_CONTAINER = 'bg-white flex flex-col lg:h-0 lg:h-screen';
 const PURPLE_SIDE_LINES_DIR = '/svgs/timeline/left_side_lines.svg';
 
-const Timeline = () => {
+const Timeline = ({ timeline }) => {
   const isEven = (index) => {
     return index % 2 == 0;
   };
@@ -21,11 +20,22 @@ const Timeline = () => {
     </div>
   );
 
+  const ImageBlock = ({ image }) => (
+    <Image
+      src={`https:${image.fields.file.url}`}
+      alt="Timeline photo"
+      placeholder="blur"
+      blurDataURL={rgbDataURL(175, 175, 175)}
+      layout="fill"
+      objectFit="cover"
+    />
+  );
+
   const EvenTimelineBlock = ({ year, description, smallImage, largeImage }) => (
     <div className="flex flex-row flex-start">
       <div className="flex flex-col">
         <div className="w-[29vw] mx-[10vw] lg:w-[14.5vw] lg:mx-[5vw] h-32 lg:h-40 mt-10 rounded-xl overflow-hidden drop-shadow-md relative">
-          <Image src={smallImage} alt="Timeline photo" placeholder="blur" layout="fill" objectFit="cover" />
+          <ImageBlock image={smallImage} />
         </div>
         <div className="flex flex-row ml-auto mt-10">
           <div className="px-3 text-5xl font-semibold">{year}</div>
@@ -35,7 +45,7 @@ const Timeline = () => {
       </div>
       {MiddleLine}
       <div className="w-[43vw] ml-[7vw] lg:w-[21.5vw] lg:ml-[3.5vw] my-16 rounded-l-xl overflow-hidden drop-shadow-md relative">
-        <Image src={largeImage} alt="Timeline photo" placeholder="blur" layout="fill" objectFit="cover" />
+        <ImageBlock image={largeImage} />
       </div>
     </div>
   );
@@ -43,12 +53,12 @@ const Timeline = () => {
   const OddTimelineBlock = ({ year, description, smallImage, largeImage }) => (
     <div className="flex flex-row flex-start">
       <div className="w-[42vw] mr-[7vw] lg:w-[21vw] lg:mr-[3.5vw] my-16 rounded-r-xl overflow-hidden drop-shadow-md relative">
-        <Image src={smallImage} alt="Timeline photo" placeholder="blur" layout="fill" objectFit="cover" />
+        <ImageBlock image={smallImage} />
       </div>
       {MiddleLine}
       <div className="flex flex-col">
         <div className="w-[30vw] mx-[10vw] h-32 lg:h-40 lg:w-[15vw] lg:mx-[5vw] mt-10 rounded-xl overflow-hidden drop-shadow-md relative">
-          <Image src={largeImage} alt="Timeline photo" placeholder="blur" layout="fill" objectFit="cover" />
+          <ImageBlock image={largeImage} />
         </div>
         <div className="flex flex-row mr-auto mt-10">
           <div className="top-1/2 relative border-t-4 border-[#7055FD] w-10" />
@@ -59,7 +69,7 @@ const Timeline = () => {
     </div>
   );
 
-  const renderTimeline = TIMELINE_DATA.reverse().map((props, index) => {
+  const renderTimeline = timeline.reverse().map((props, index) => {
     return isEven(index) ? EvenTimelineBlock(props) : OddTimelineBlock(props);
   });
